[3d_viewer] Added a 'unique color' rendering mode for object picking and visibility check
In this mode, every single object is rendered with a different color,
and a look-up table is used to retrieve which nodes have been seen or
are hovered by the mouse.

This commit takes also advantage to this change to implement an simple
object selection (click on it) + displacement (with arrows).

By pressing 'v', the list of current visible objects is output.

While here, added a simple screenshot function when pressing 's'.