protobuf support for mesh loading -- 'collada' test pass
The test sometimes fails, sometimes pass. It seems to be due to some invalidation requests being missed
when the server is busy receiving the mesh.

Mesh transfer is *slow* (~12 sec for the monkeys test scene). Serialization is slow as well (~1sec for 15500 faces).
Better encoding of the mesh would certainly help.