Major rework of the update mechanism for improved performances
The clients now embed an 'invalidation server' to which the main uwds server connect to push
worlds' update. This remove the need for a polling thread on the client side.

Various other changes have been performed to improve the overall performance, and after
extensive testing, it appears that most of the time is waisted on waiting the the
'waitforchanges' condition variable to be notified (like ~400ms out of 500ms).

When using Python3, the issue almost completely disappear and the
performance is much better (~80ms to propagate a change across 10 worlds)