Allow update/delete of multiple nodes/situations
*This commit changes underworlds proto*, however the
API is backward compatible.

append()/update()/remove() nodes or situations now accept either
a single node/situation ID (as previously) or a list of IDs.

In the later case, the IDs are transmitted in a single batch, saving
round-trips with the server.

The invalidations can now be sent in batch. However, this is not yet
taken advantage of by the server.

All existing unit-tests pass, and testing/nodes.py has been sligthly
extended to test bulk addition & deletions.